+++ ai-agents/crowd-fund-analysis/cf_analysis_agent/controller.py
@@ -3,6 +3,42 @@ import argparse
 from agent import graph as parent_graph
 from cf_analysis_agent.agent_state import AgentState, ProjectInfo
 from cf_analysis_agent.utils.project_utils import scrape_urls
+from cf_analysis_agent.utils.report_utils import get_project_info_from_s3
+
+
+def prepare_processing_command(project_id, model, script_path="cf_analysis_agent/controller.py"):
+    """
+    Prepares the command to start processing based on variables extracted from S3.
+
+    Args:
+        project_id (str): The project ID to extract variables for.
+        model (str): The selected AI model for report regeneration.
+        script_path (str): Path to the script to be executed. Defaults to "cf_analysis_agent/controller.py".
+
+    Returns:
+        list: The prepared command as a list of arguments.
+    """
+    # Extract variables from S3
+    variables: ProjectInfo = get_project_info_from_s3(project_id)
+
+    # Base command
+    command = [
+        "poetry", "run", "python", script_path,
+        variables["project_id"],
+        variables["project_name"],
+        variables["crowdfunding_link"],
+        variables["website_url"],
+        variables["latest_sec_filing_link"]
+    ]
+
+    # Include additional links if they exist
+    if variables.get("additional_links"):
+        command.extend(["--additional_links", ",".join(variables["additional_links"])])
+
+    # Append the selected model as an argument
+    command.extend(["--model", model])
+
+    return command
 
 
 def parse_arguments() -> AgentState:
@@ -10,6 +46,7 @@ def parse_arguments() -> AgentState:
     Parses command-line arguments and returns project details.
     """
     parser = argparse.ArgumentParser(description="Run the project report generator.")
+    parser.add_argument("project_id", help="The id of the project.")
     parser.add_argument("project_name", help="The name of the project.")
     parser.add_argument("crowdfunding_link", help="The crowdfunding link for the project.")
     parser.add_argument("website_url", help="The official website URL of the project.")
@@ -33,6 +70,7 @@ def parse_arguments() -> AgentState:
     args = parser.parse_args()
 
     # Sanitize inputs to remove unnecessary quotes or whitespace
+    project_id = args.project_id.strip().strip('"')
     project_name = args.project_name.strip().strip('"')
     crowdfunding_link = args.crowdfunding_link.strip().strip('"')
     website_url = args.website_url.strip().strip('"')
@@ -41,16 +79,15 @@ def parse_arguments() -> AgentState:
     report_type = args.report_type.strip().strip('"') if args.report_type else "all"
     model = args.model.strip().strip('"') if args.model else None
 
-    project_id = project_name.replace(" ", "_").lower()
-
     project_info: ProjectInfo = {
+        "project_id": project_id,
         "project_name": project_name,
         "crowdfunding_link": crowdfunding_link,
         "website_url": website_url,
         "latest_sec_filing_link": latest_sec_filing_link,
         "additional_links": additional_links,
-        "project_id": project_id,
     }
+
     scraped_urls = scrape_urls(list([crowdfunding_link, website_url]))
     combined_content = "\n\n".join(scraped_urls)
     return {
@@ -65,8 +102,7 @@ def parse_arguments() -> AgentState:
         "reports_to_generate": None,
         "processed_project_info": {
             "combined_scrapped_content": combined_content
-        },
-        "general_info_report": None
+        }
     }
 
 
@@ -81,3 +117,5 @@ async def main_controller_async(agent_state: AgentState):
 if __name__ == "__main__":
     initialState = parse_arguments()
     asyncio.run(main_controller_async(initialState))
+
+
